Rename conditions2 to conditions
