+++ pkg/conditions/getter.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package conditions2
+package conditions
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
