+++ pkg/conditions/merge.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package conditions2
+package conditions
 
 import (
 	"sort"
