+++ pkg/prober/worker/readiness_worker_test.go
@@ -19,7 +19,7 @@ import (
 
 	vmopv1 "github.com/vmware-tanzu/vm-operator/api/v1alpha2"
 
-	conditions "github.com/vmware-tanzu/vm-operator/pkg/conditions2"
+	"github.com/vmware-tanzu/vm-operator/pkg/conditions"
 	"github.com/vmware-tanzu/vm-operator/pkg/prober/context"
 	fakeprobe "github.com/vmware-tanzu/vm-operator/pkg/prober/fake/probe"
 	"github.com/vmware-tanzu/vm-operator/pkg/prober/probe"
